+++ sys/arch/mips/include/mips_param.h
@@ -107,7 +107,7 @@
 
 #ifdef _LP64
 #define	NSEGPG		(NBPG/8)
-#define NBXSEG		(NSEGPG*NBSEG)	/* bytes/xsegment */
+#define NBXSEG		((uint64_t)NSEGPG*NBSEG) /* bytes/xsegment */
 #define	XSEGOFSET	(NBSEG-1)	/* byte offset into segment */
 #define	XSEGSHIFT	(SEGSHIFT+(PGSHIFT-3))	/* LOG2(NBXSEG) */
 #endif
